requirements.txt
Django==2.0.13
pytz==2019.1
PyJWT==1.7.1